adding login and edit article + bug fixes

+++ src/app/doctor-details/doctor-details.component.ts
@@ -27,41 +27,45 @@ export class DoctorDetailsComponent implements OnInit {
 
   getDoctorDetails() {
     return this.doctorService.getDoctotId()
-    .pipe(switchMap(id => {
-      return forkJoin([this.doctorService.getDoctorDetails(id),
+      .pipe(switchMap(id => {
+        return forkJoin([this.doctorService.getDoctorDetails(id),
         this.doctorService.getDoctorSchedule(id),
         this.doctorService.getDoctorPreviousAppointments(id),
         this.doctorService.getDoctorUpcomingAppointments(id)
-      ])
-    }))
-    .subscribe(
-      (data) => {
-        this.doctor = data[0];
-        this.doctor.schedule = data[1];
-        this.doctor.previousAppointments = data[2];
-        this.doctor.upcomingAppointments = data[3];
-        
-        this.loading = false;
-      }
-    );
+        ])
+      }))
+      .subscribe(
+        (data) => {
+          this.doctor = data[0];
+          this.doctor.schedule = data[1];
+          this.doctor.previousAppointments = data[2];
+          this.doctor.upcomingAppointments = data[3];
+
+          this.loading = false;
+        }
+      );
   }
 
-  deleteDoctor(){
-    return this.doctorService.getDoctotId()
-    .pipe(switchMap(id => {
-      return forkJoin([this.doctorService.deleteDoctor(id)])
-    }))
-    .subscribe(
-      (data:any) => {
-        this.deleteSuccess = data[0].isSuccess;
-        this.msg = data[0].message;
+  deleteDoctor() {
+    if (confirm("Are you sure to delete this doctor?")) {
+      return this.doctorService.getDoctotId()
+        .pipe(switchMap(id => {
+          return forkJoin([this.doctorService.deleteDoctor(id)])
+        }))
+        .subscribe(
+          (data: any) => {
+            this.deleteSuccess = data[0].isSuccess;
+            this.msg = data[0].message;
 
-        this.delay(3000).then(()=>this.router.navigate(['/']));
-      }
-    );
+            this.delay(3000).then(() => this.router.navigate(['/']));
+          }
+        );
+    } else {
+      return;
+    }
   }
 
   delay(ms: number) {
-    return new Promise( resolve => setTimeout(resolve, ms) );
-}
+    return new Promise(resolve => setTimeout(resolve, ms));
+  }
 }
